fix: left click to leave queueing back
Previously, left click to leave wouldn't prevent the actual click from
being sent to the server. Because of that sometimes the server would
process /leave first, and then the click, so the player would be put
back in queue. This commit fixes it by cancelling the left click if the
event consumers say so.

Also makes left click update the temporary queue sign block.
This makes both left click to leave and explicit leave less buggy.